requirements.txt
django-filer
django-admin-search
django-tinymce